# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Guised Up API
  version: 1.0.0
  description: |
    Real Connections Feed API — a social platform backend that ranks content by semantic
    relevance, relationship strength, authenticity score, and time decay rather than
    simple chronological order or raw engagement.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained from `/api/auth/login` or
    `/api/auth/register`. Pass it as `Authorization: Bearer <accessToken>`.

    ## Rate Limiting
    All `/api/` routes are rate-limited to **200 requests per 15 minutes** per IP.
    Exceeding this returns HTTP 429.
servers:
  - url: http://localhost:3000
    description: Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login or /api/auth/register

  # ───────────────────────────────────────────────────────────────────────────
  # Reusable schemas
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Something went wrong"
    ValidationError:
      allOf:
        - $ref: "#/components/schemas/ErrorEnvelope"
        - type: object
          properties:
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
              example:
                - field: email
                  message: Invalid email address

    # ── Auth ─────────────────────────────────────────────────────────────────
    UserPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        email:
          type: string
          format: email
          example: "alice@example.com"
        username:
          type: string
          example: "alice"
        avatarUrl:
          type: string
          nullable: true
          example: "https://cdn.example.com/avatars/alice.jpg"
        bio:
          type: string
          nullable: true
          example: "Building real connections."
    AuthorRef:
      type: object
      description: Lightweight author reference embedded in posts and comments
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        username:
          type: string
          example: "alice"
        avatarUrl:
          type: string
          nullable: true
          example: "https://cdn.example.com/avatars/alice.jpg"
...
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "p1b2c3d4-e5f6-7890-abcd-ef1234567890"
        authorId:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        text:
          type: string
          example: "Just had a genuine conversation with a stranger on the bus. Highly recommend."
        imageUrl:
          type: string
          nullable: true
          example: "https://cdn.example.com/posts/img123.jpg"
        authenticityScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: ML-derived authenticity score for this post
          example: 0.87
        viewCount:
          type: integer
          example: 142
        createdAt:
          type: string
          format: date-time
          example: "2026-06-15T09:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-06-15T09:30:00.000Z"
        author:
          $ref: "#/components/schemas/AuthorRef"
        _count:
          type: object
          properties:
            interactions:
              type: integer
              example: 23
...
    PostSingleResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Post"
    CreatePostBody:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 2000
          example: "Just had a genuine conversation with a stranger on the bus."
        imageUrl:
          type: string
          nullable: true
          example: "https://cdn.example.com/posts/img123.jpg"
    UpdatePostBody:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 2000
          example: "Updated post text."
        imageUrl:
          type: string
          nullable: true
          description: Pass null to remove the image
          example: null

    # ── Feed ─────────────────────────────────────────────────────────────────
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "c1b2c3d4-e5f6-7890-abcd-ef1234567890"
        postId:
          type: string
          format: uuid
          example: "p1b2c3d4-e5f6-7890-abcd-ef1234567890"
        authorId:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        text:
          type: string
          example: "This really resonated with me!"
        createdAt:
          type: string
          format: date-time
          example: "2026-06-15T10:05:00.000Z"
        author:
          $ref: "#/components/schemas/AuthorRef"
    CommentsListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
    CreateCommentBody:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1000
          example: "This really resonated with me!"

    # ── Notifications ────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: "Unauthorized"
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: "Forbidden"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: "Not found"
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: "Too many requests, please try again later"
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: "Internal server error"

  # ───────────────────────────────────────────────────────────────────────────
  # Reusable parameters
  # ───────────────────────────────────────────────────────────────────────────
  parameters:
    PostIdPath:
      name: id
      in: path
      required: true
      description: Post UUID
      schema:
        type: string
        format: uuid
      example: "p1b2c3d4-e5f6-7890-abcd-ef1234567890"
    PostIdInComments:
      name: postId
      in: path
      required: true
      description: Post UUID (parent of comments)
      schema:
        type: string
        format: uuid
      example: "p1b2c3d4-e5f6-7890-abcd-ef1234567890"
    CommentIdPath:
      name: id
      in: path
      required: true
      description: Comment UUID
      schema:
        type: string
        format: uuid
      example: "c1b2c3d4-e5f6-7890-abcd-ef1234567890"
...
paths:
...
  /api/posts:
    post:
      tags: [posts]
      operationId: createPost
      summary: Create a new post
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostSingleResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/posts/{id}:
    get:
      tags: [posts]
      operationId: getPostById
      summary: Get a single post by ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdPath"
      responses:
        "200":
          description: Post found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostSingleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags: [posts]
      operationId: updatePost
      summary: Update a post (owner only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostSingleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags: [posts]
      operationId: deletePost
      summary: Delete a post (owner only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdPath"
      responses:
        "200":
          description: Post deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── Feed ───────────────────────────────────────────────────────────────────
...
  /api/posts/{postId}/comments:
    get:
      tags: [comments]
      operationId: getComments
      summary: List comments on a post
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdInComments"
      responses:
        "200":
          description: Comments list (chronological)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentsListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [comments]
      operationId: createComment
      summary: Add a comment to a post
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdInComments"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Comment"
        "400":
          description: Missing text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/posts/{postId}/comments/{id}:
    delete:
      tags: [comments]
      operationId: deleteComment
      summary: Delete a comment (owner or post-owner only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdInComments"
        - $ref: "#/components/parameters/CommentIdPath"
      responses:
        "200":
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── Notifications ──────────────────────────────────────────────────────────
...
  /api/users/search:
    get:
      tags: [users]
      operationId: searchUsers
      summary: Search for users by username or display name
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Username prefix or display name fragment to search for
          schema:
            type: string
            minLength: 1
          example: "ali"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
          example: 10
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/UserPublic"
        "400":
          description: Missing `q` parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...